Align file names to namespace
