+++ src/Extensions/AindBehaviorVrForaging.cs
@@ -1,26 +1,99 @@
-﻿using Bonsai;
 using System;
+using System.Collections.Generic;
 using System.Linq;
+using MathNet.Numerics.Interpolation;
+using Bonsai;
 using System.Reactive.Linq;
 using MathNet.Numerics.Distributions;
 using System.ComponentModel;
 
+
 namespace AindVrForagingDataSchema
 {
+    public partial class PatchUpdateFunction
+    {
+        public readonly Random defaultRandom = new Random();
+        public virtual double Invoke(double value, double tickValue, Random random = null)
+        {
+            throw new NotImplementedException();
+        }
+    }
+
+    public partial class SetValueFunction
+    {
+        public override double Invoke(double value, double tickValue, Random random = null)
+        {
+            if (random == null) random = defaultRandom;
+            return this.Value.SampleDistribution(random);
+        }
+    }
+
+    public partial class ClampedRateFunction
+    {
+        public override double Invoke(double value, double tickValue, Random random = null)
+        {
+            if (random == null) random = defaultRandom;
+            double newValue = value + this.Rate.SampleDistribution(random) * tickValue;
+
+            newValue = Math.Min(this.Maximum.HasValue ? this.Maximum.Value : newValue, newValue);
+            newValue = Math.Max(this.Minimum.HasValue ? this.Minimum.Value : newValue, newValue);
+            return newValue;
+        }
+    }
+
+    public partial class ClampedMultiplicativeRateFunction
+    {
+        public override double Invoke(double value, double tickValue, Random random = null)
+        {
+            if (random == null) random = defaultRandom;
+            double newValue = value * Math.Pow(this.Rate.SampleDistribution(random), tickValue);
+
+            newValue = Math.Min(this.Maximum.HasValue ? this.Maximum.Value : newValue, newValue);
+            newValue = Math.Max(this.Minimum.HasValue ? this.Minimum.Value : newValue, newValue);
+            return newValue;
+        }
+    }
+
+    public partial class LookupTableFunction
+    {
+        private Dictionary<double, double> ToLookupTable()
+        {
+            return LutKeys.Zip(LutValues, (key, value) => new { Key = key, Value = value })
+                         .ToDictionary(x => x.Key, x => x.Value);
+        }
+
+        public override double Invoke(double value, double tickValue, Random random = null)
+        {
+            Dictionary<double, double> Table = ToLookupTable();
+            if (value > Table.Keys.Max())
+            {
+                return Table[Table.Keys.Max()];
+            }
+            if (value < Table.Keys.Min())
+            {
+                return Table[Table.Keys.Min()];
+            }
+            IInterpolation interpolation = MathNet.Numerics.Interpolate.Linear(Table.Keys, Table.Values);
+            return interpolation.Interpolate(value);
+        }
+    }
 
     partial class Distribution
     {
         private const uint SampleSize = 1000;
 
-        public virtual double SampleDistribution(Random random){
+        public virtual double SampleDistribution(Random random)
+        {
             throw new NotImplementedException();
         }
 
-        public virtual IDiscreteDistribution GetDiscreteDistribution(Random random){
+        public virtual IDiscreteDistribution GetDiscreteDistribution(Random random)
+        {
             throw new NotImplementedException();
         }
 
-        public virtual IContinuousDistribution GetContinuousDistribution(Random random){
+        public virtual IContinuousDistribution GetContinuousDistribution(Random random)
+        {
             throw new NotImplementedException();
         }
 
@@ -31,7 +104,8 @@ namespace AindVrForagingDataSchema
 
         public double DrawSample(IContinuousDistribution distribution, ScalingParameters scalingParameters, TruncationParameters truncationParameters)
         {
-            if (truncationParameters != null){
+            if (truncationParameters != null)
+            {
                 ValidateTruncationParameters(truncationParameters);
                 var samples = SampleFromDistribution(distribution, truncationParameters, scalingParameters);
                 return ValidateSamples(samples.Item1, samples.Item2, truncationParameters);
@@ -44,7 +118,8 @@ namespace AindVrForagingDataSchema
 
         public double DrawSample(IDiscreteDistribution distribution, ScalingParameters scalingParameters, TruncationParameters truncationParameters)
         {
-            if (truncationParameters != null){
+            if (truncationParameters != null)
+            {
                 ValidateTruncationParameters(truncationParameters);
                 var samples = SampleFromDistribution(distribution, truncationParameters, scalingParameters);
                 return ValidateSamples(samples.Item1, samples.Item2, truncationParameters);
@@ -93,8 +168,10 @@ namespace AindVrForagingDataSchema
                     throw new ArgumentException("Truncation heuristic has failed. Please check your truncation parameters.");
                 }
             }
-            else {
-                outValue = drawnSamples.First(); }
+            else
+            {
+                outValue = drawnSamples.First();
+            }
             return outValue;
         }
 
@@ -107,8 +184,10 @@ namespace AindVrForagingDataSchema
         }
     }
 
-    partial class PdfDistribution{
-        public override double SampleDistribution(Random random){
+    partial class PdfDistribution
+    {
+        public override double SampleDistribution(Random random)
+        {
             var pdf = DistributionParameters.Pdf;
             var index = DistributionParameters.Index;
             if (pdf.Count != index.Count)
@@ -130,111 +209,129 @@ namespace AindVrForagingDataSchema
         }
     }
 
-    partial class Scalar{
-        public override double SampleDistribution(Random random){
+    partial class Scalar
+    {
+        public override double SampleDistribution(Random random)
+        {
             return DistributionParameters.Value;
         }
     }
 
-    partial class NormalDistribution{
+    partial class NormalDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new Normal(DistributionParameters.Mean, DistributionParameters.Std, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class ExponentialDistribution{
+    partial class ExponentialDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new Exponential(DistributionParameters.Rate, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class LogNormalDistribution{
+    partial class LogNormalDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new LogNormal(DistributionParameters.Mean, DistributionParameters.Std, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class GammaDistribution{
+    partial class GammaDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new Gamma(DistributionParameters.Shape, DistributionParameters.Rate, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class BetaDistribution{
+    partial class BetaDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new Beta(DistributionParameters.Alpha, DistributionParameters.Beta, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class UniformDistribution{
+    partial class UniformDistribution
+    {
 
         public override IContinuousDistribution GetContinuousDistribution(Random random)
         {
             return new ContinuousUniform(DistributionParameters.Min, DistributionParameters.Max, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetContinuousDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class BinomialDistribution{
+    partial class BinomialDistribution
+    {
 
         public override IDiscreteDistribution GetDiscreteDistribution(Random random)
         {
             return new Binomial(DistributionParameters.P, DistributionParameters.N, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetDiscreteDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
     }
 
-    partial class PoissonDistribution{
+    partial class PoissonDistribution
+    {
 
         public override IDiscreteDistribution GetDiscreteDistribution(Random random)
         {
             return new Poisson(DistributionParameters.Rate, random);
         }
 
-        public override double SampleDistribution(Random random){
+        public override double SampleDistribution(Random random)
+        {
             var distribution = GetDiscreteDistribution(random);
             return DrawSample(distribution, ScalingParameters, TruncationParameters);
         }
